update(processor): cast processed data to float

+++ workflows/airflow/dags/modules/processor/__init__.py
@@ -11,8 +11,8 @@ TRAIN_TEST_SPLIT = 0.2
 def pre_data_preparation(csv_string):
     df_diamonds = pd.read_csv(io.StringIO(csv_string))
 
-    # the analysis didn't reveal any missing data
-    # however, it's recommended to still consider removing incomplete data
+    # The analysis didn't reveal any missing data.
+    # However, it's recommended to still consider removing incomplete data.
 
     df_diamonds = df_diamonds[
         (df_diamonds.x * df_diamonds.y * df_diamonds.z != 0)
@@ -33,7 +33,7 @@ def linear_data_preparation(diamonds):
         df_diamonds,
         columns=['cut', 'color', 'clarity'],
         drop_first=True,
-        dtype=np.float32
+        dtype=float
     )
 
     return [df_diamonds.columns.tolist()] + df_diamonds.values.tolist()
@@ -65,6 +65,8 @@ def xgboost_data_preparation(diamonds):
 
 def post_data_preparation(diamonds):
     df_diamonds = pd.DataFrame(diamonds[1:], columns=diamonds[0])
+
+    df_diamonds = df_diamonds.astype(float)
 
     x = np.array(df_diamonds.drop(columns='price'))
     y = np.log(np.array(df_diamonds.price, dtype=float))
